Pull request 2- DEV to MAIN

+++ code/remove_data_from_s3.py
@@ -1,6 +1,8 @@
-from utils import connect_to_aws_service_client, load_config
+from utils import connect_to_aws_service_client, load_config, create_spark_session, arg_parser
 import glob
 import os 
+
+APP_NAME = "remove_data_from_s3"
 
 def delete_bucket(bucket_name):
     s3_client.delete_bucket(Bucket=bucket_name)
@@ -9,6 +11,7 @@ def delete_data_from_bucket(bucket_name, key):
     s3_client.delete_object(Bucket=bucket_name, Key=key)
 
 if __name__ == '__main__':
+    spark = create_spark_session(APP_NAME)
 
     config_data = load_config() 
 
@@ -23,6 +26,8 @@ if __name__ == '__main__':
     
     s3_client = connect_to_aws_service_client('s3', access_key, secret_key, region_name)
 
+    ngrams = arg_parser('Please state cleanup or delete')
+
     try:
         for obj_list in s3_client.list_objects(Bucket=bucket_name)['Contents']:
             key = obj_list['Key']
@@ -31,7 +36,8 @@ if __name__ == '__main__':
     except:
         print('Bucket is empty. Nothing to delete.')
     
-    try:
-        delete_bucket(bucket_name)
-    except:
-        print('Bucket does not exist.')
+    if ngrams == 'delete': 
+        try:
+            delete_bucket(bucket_name)
+        except:
+            print('Bucket does not exist.')
